### 2024-03-22 #6529 @editorialbot
[PRE REVIEW]: ICAT: The Interactive Corpus Analysis Tool
**Submitting author:** @WarmCyan (<a href="http://orcid.org/0000-0002-5036-5433">Nathan Martindale</a>)
**Repository:** https://github.com/ORNL/icat
**Branch with paper.md** (empty if default branch):
**Version:** v0.7.2
