### 2024-05-01 #6529 @WarmCyan
Hello @JBorrow! In two previous JOSS submissions we've had deniederhut (github username) as a reviewer, if he's not tired of looking through our stuff yet and has time/interest, he's always provided really useful feedback previously! Other than that, no particular suggestions
