### 2024-05-08 #6529 @JBorrow
Hi @deniederhut, given your experience with the author's previous work, we'd like to invite you to review this submission for JOSS too. Please let me know if you are interested!
