### 2024-05-08 #6529 @JBorrow
Hi @thomaskrause, I found your name in the JOSS reviewers database. Would you be interested in reviewing this submission for us? More information on reviewing for JOSS is available in the docs: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.
