### 2024-06-06 #6529 @JBorrow
Hi @jhagerer and @KennethEnevoldsen, I found your names in the JOSS reviewers database. Would you be interested in reviewing this submission for us? More information on reviewing for JOSS is available in the docs: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.
