### 2024-06-06 #6529 @KennethEnevoldsen
Hi @JBorrow, thanks for reaching out. I sadly do not have the time atm. (I have a thesis deadline incoming), so will sadly have to decline. The work looks both interesting and relevant though.
