### 2024-06-09 ORNL/icat#44 @jhagerer
JOSS review: Runtime errors after setting up an environment
Dear @WarmCyan, I am the reviewer of your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/6529).

I run into problems after setting up the environment. When running `icat.initialize(offline=True)`, the following errors occur:

![image](https://github.com/ORNL/icat/assets/875050/11c76a09-e01f-456a-b4c3-be10ed4f5432)

One way to reproduce:

```
git clone https://github.com/ORNL/icat.git
cd icat
conda env create --file environment.yml  # to set the right python version in a fresh virtual environment
conda activate icat
conda install pip
pip install icat-iml
conda install jupyter_bokeh  # necessary after ModuleNotFoundError
pip install ipywidgets-bokeh  # necessary after ModuleNotFoundError
pip install -r requirements.txt  # tried that for debugging, didn't help
```

As a first suggestion, you might want to add `ipywidgets-bokeh`to requirements.txt and `jupyter_bokeh` to environment.yml (or both to requirements.txt). Secondly, I would recommend to run the setup by yourself from scratch the way I did it, which is quite close to your [documentation](https://ornl.github.io/icat/stable/user/getting_started.html). Alternatively, you might want to give me feedback where I made something wrong during the setup.

I put the JOSS submission review on hold and try again next weekend. Please tag me here or in the [review thread](https://github.com/openjournals/joss-reviews/issues/6529) if you have updates or suggestions. Thanks!
