### 2024-06-09 #6529 @jhagerer
@JBorrow @WarmCyan I ran into problems while setting up the project, see [this ticket](https://github.com/ORNL/icat/issues/44). I have to put the review on hold until we have clarification on this. After I receive updates, I can proceed next weekend.
