### 2024-06-09 #6529 @jhagerer
Current state of the checklist, which will be updated later:

* [x]  **Repository:** Is the source code for this software available at the https://github.com/KennethEnevoldsen/augmenty?
