### 2024-06-10 #6529 @JBorrow
Hi @jhagerer, thank you for this! We're currently in the pre-review stage, and the actual review will take place in another thread; I presume given this you are happy to review the package (given the issue will be fixed and you can continue)? If so, I can assign you as a reviewer and we will move to a separate thread.
