### 2024-06-11 #6873 @editorialbot
[REVIEW]: ICAT: The Interactive Corpus Analysis Tool
**Submitting author:** @WarmCyan (<a href="http://orcid.org/0000-0002-5036-5433">Nathan Martindale</a>)
**Repository:** https://github.com/ORNL/icat
**Branch with paper.md** (empty if default branch):
**Version:** v0.8.1
