### 2024-06-11 #6873 @JBorrow
Hi @WarmCyan! Would you please be able to address the missing DOIs in the references?
