### 2024-06-12 ORNL/icat#44 @WarmCyan
Hey @jhagerer! Thanks for agreeing to review this!

Just a quick update, I started taking a look at this yesterday, go figure in the couple months since I last tested setting this up from scratch, various package updates have somehow broken it :weary:

I don't _think_ either `ipywidgets-bokeh` or `jupyter_bokeh` should strictly be necessary for running this in JupyterLab, or at least in my (working) development environment I don't have either installed. As far as I'm aware, those libraries are used depending on what context you're actually running stuff in (putting this here for my own future reference: https://panel.holoviz.org/explanation/comms.html#libraries-modules-with-support-for-communication-in-these-contexts), but you raise a good point that it'd probably be better to just go ahead and include them to maximally cover various contexts. Moving forward I'll add them into the setup.

(Notably I haven't actually tried running this from vscode before. I know every now and then that can do odd things with widgets, but given it's not even working in JupyterLab in a new environment either, that shouldn't be the issue here)

I'm still working to figure out the most recent magical combination of version numbers that work (I think the issue is somewhere in the interplay between bokeh, panel, and the recent pyviz>3). If I don't get something satisfactory by the weekend, I'll check that the exact versions I have in my development environment work in a new env and list those here so you at least have something to play with.
