### 2024-06-13 165aaff
Add panel widget comms dependencies and browser-based playwright test
