+++ requirements.txt
@@ -14,3 +14,4 @@ pre-commit
 sphinx~=7.1.2
 pydata-sphinx-theme
 sphinx-favicon
+pytest-playwright
