### 2024-06-13 d7c4d51
Update makefile style
