### 2024-06-13 3993e5d
Add playwright install step to github tests
