+++ .github/workflows/tests.yml
@@ -23,5 +23,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
+          playwright install
       - name: Test with pytest
         run: pytest -vv
