+++ requirements.txt
@@ -15,3 +15,4 @@ sphinx~=7.1.2
 pydata-sphinx-theme
 sphinx-favicon
 pytest-playwright
+jupyterlab
