### 2024-06-13 96d50bb
Change browser test to use 127.0.0.1 instead of localhost
