### 2024-06-13 ff5b4c6
Try adding pause before playwright starts
