### 2024-06-13 cba558b
Improve jupyter readiness check in browser test
