+++ .github/workflows/tests.yml
@@ -25,4 +25,4 @@ jobs:
           pip install -r requirements.txt
           playwright install
       - name: Test with pytest
-        run: pytest -vv
+        run: pytest -vv -s
