### 2024-06-14 a8420e2
Add extremely restrictive setup dependencies to get back to a working instance
