+++ setup.py
@@ -40,17 +40,36 @@ setup(
     ],
     python_requires=">=3.10",
     packages=["icat"],
+    # install_requires=[
+    #     "panel",
+    #     "pyviz_comms",
+    #     "numpy",
+    #     "pandas",
+    #     "scikit-learn",
+    #     "ipyanchorviz",
+    #     "altair",
+    #     #"jupyter_bokeh",
+    #     #"ipywidgets-bokeh",
+    #     "ipyvuetify",
+    #     "ipywidgets==8.1.2",
+    #     "jupyterlab-widgets==3.0.10",
+    # ],
     install_requires=[
-        "panel",
+        "panel<1.2",
         "pyviz_comms<3",
         "numpy",
         "pandas",
         "scikit-learn",
-        "ipyvuetify",
-        "ipywidgets",
         "ipyanchorviz",
         "altair",
-        "jupyter_bokeh",
-        "ipywidgets-bokeh",
+        "ipyvue==1.9.1",
+        "ipyvuetify==1.8.10",
+        "ipywidgets==8.0.6",
+        "param<2",
+        "traitlets==5.9.0",
+        "bokeh==3.1.1",
+        "jupyterlab==4.1.2",
+        "jupyterlab_server==2.23.0",
+        "jupyterlab-widgets==3.0.7",
     ],
 )
