### 2024-06-14 9267b8e
Add more time to browser check
