### 2024-06-17 5d74877
Add ipywidgets-bokeh etc to restrictive setup
