+++ paper/paper.bib
@@ -7,6 +7,7 @@
   eprint = {1707.06742},
   eprinttype = {arxiv},
   eprintclass = {cs, stat},
+  doi = {10.48550/arXiv.1707.06742},
   url = {http://arxiv.org/abs/1707.06742},
   urldate = {2022-04-13},
   abstract = {The current processes for building machine learning systems require practitioners with deep knowledge of machine learning. This significantly limits the number of machine learning systems that can be created and has led to a mismatch between the demand for machine learning systems and the ability for organizations to build them. We believe that in order to meet this growing demand for machine learning systems we must significantly increase the number of individuals that can teach machines. We postulate that we can achieve this goal by making the process of teaching machines easy, fast and above all, universally accessible. While machine learning focuses on creating new algorithms and improving the accuracy of "learners", the machine teaching discipline focuses on the efficacy of the "teachers". Machine teaching as a discipline is a paradigm shift that follows and extends principles of software engineering and programming languages. We put a strong emphasis on the teacher and the teacher's interaction with data, as well as crucial components such as techniques and design principles of interaction and visualization. In this paper, we present our position regarding the discipline of machine teaching and articulate fundamental machine teaching principles. We also describe how, by decoupling knowledge about machine learning algorithms from the process of teaching, we can accelerate innovation and empower millions of new uses for machine learning models.},
@@ -106,6 +107,7 @@
 @misc{panelholoviz,
 title={Panel: The powerful data exploration \& web app framework for Python},
 author={Holoviz},
+doi = {10.5281/zenodo.3706648},
 howpublished={\url{https://panel.holoviz.org/}},
 year={2018},
 note={Accessed: 2023-10-10},
@@ -125,6 +127,7 @@ note={Accessed: 2023-10-12},
          and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
          and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
          Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+ doi = {10.48550/arXiv.1201.0490}
  journal={Journal of Machine Learning Research},
  volume={12},
  pages={2825--2830},
