### 2024-06-17 #6873 @WarmCyan
Hello @JBorrow! I've added DOIs for all except ipyvuetify, I couldn't find anything for that one
