### 2024-06-17 ORNL/icat#44 @WarmCyan
Still continuing to work on this, sorry I didn't get anything to you by the weekend, I've been jumping from fire to fire!

I've figured out at least one of the issues, I think there's some sort of a breaking change between the latest ipywidgets 8.1.3 version and panel when used together in JupyterLab, I've separately opened up an issue https://github.com/holoviz/panel/issues/6921 because I don't think it's something I can fix over here. I've also (finally) added a playwright test to automate checking that one of the notebooks actually runs and displays the interface to hopefully catch these sorts of problems sooner moving forward.

Restricting only `ipywidgets` and `jupyterlab_widgets` versions still has some issue going on that I'm trying to figure out, but for what it's worth I do now have a very strict list of dependencies in `setup.py` that sets exact versions for most relevant things in my dev environment (and nominally works, at least in JupyterLab)

In principle, to get a currently working environment with those, you should be able to run:
```bash
git clone https://github.com/ORNL/icat.git
cd icat
conda create -n icat python
conda activate icat
pip install -e .
```

I'll update again once I've narrowed down any other dependency issues and have a less restrictive setup.py
