### 2024-06-17 #6873 @WarmCyan
@SamHames thanks for agreeing to review this! Just so you're aware if you start playing with the software soon, there's a couple weird issues going on right now with dependencies that I'm still trying to resolve, if you run into problems during setup I'm tracking them over here: https://github.com/ORNL/icat/issues/44
