### 2024-06-21 c551bdd
Fix vue template issue and break vue contents out into separate files
