+++ setup.py
@@ -40,38 +40,42 @@ setup(
     ],
     python_requires=">=3.10",
     packages=["icat"],
-    # install_requires=[
-    #     "panel",
-    #     "pyviz_comms",
-    #     "numpy",
-    #     "pandas",
-    #     "scikit-learn",
-    #     "ipyanchorviz",
-    #     "altair",
-    #     "jupyter_bokeh",
-    #     "ipywidgets-bokeh",
-    #     "ipyvuetify",
-    #     "ipywidgets<8.1.3",
-    #     "jupyterlab-widgets<3.0.11",
-    # ],
+    package_data={"icat": ["vue/*"]},
     install_requires=[
-        "panel<1.2",
-        "pyviz_comms<3",
+        "panel",
+        "pyviz_comms",
         "numpy",
         "pandas",
         "scikit-learn",
         "ipyanchorviz",
         "altair",
-        "ipyvue==1.9.1",
-        "ipyvuetify==1.8.10",
-        "ipywidgets==8.0.6",
-        "param<2",
-        "traitlets==5.9.0",
-        "bokeh==3.1.1",
-        "jupyterlab==4.1.2",
-        "jupyterlab_server==2.23.0",
-        "jupyterlab-widgets==3.0.7",
-        "ipywidgets-bokeh",
         "jupyter_bokeh",
+        "ipywidgets-bokeh",
+        "ipyvuetify~=1.9",
+        "ipyvue~=1.11",
+        "ipywidgets<8.1.3",  # can be removed once https://github.com/holoviz/panel/issues/6921 is resolved (panel 1.5?)
+        "jupyterlab-widgets<3.0.11",  # "
     ],
+    # an exact set of dependencies that I know for a fact works, remove comments
+    # in case of emergency
+    # install_requires=[
+    #     "panel<1.2",
+    #     "pyviz_comms<3",
+    #     "numpy",
+    #     "pandas",
+    #     "scikit-learn",
+    #     "ipyanchorviz",
+    #     "altair",
+    #     "ipyvue==1.9.1",
+    #     "ipyvuetify==1.8.10",
+    #     "ipywidgets==8.0.6",
+    #     "param<2",
+    #     "traitlets==5.9.0",
+    #     "bokeh==3.1.1",
+    #     "jupyterlab==4.1.2",
+    #     "jupyterlab_server==2.23.0",
+    #     "jupyterlab-widgets==3.0.7",
+    #     "ipywidgets-bokeh",
+    #     "jupyter_bokeh",
+    # ],
 )
