+++ icat/table.py
@@ -5,7 +5,7 @@ from collections.abc import Callable
 import ipyvuetify as v
 import traitlets
 
-from icat.utils import _kill_param_auto_docstring
+from icat.utils import _kill_param_auto_docstring, vue_template_path
 
 _kill_param_auto_docstring()
 
@@ -14,6 +14,8 @@ class TableContentsTemplate(v.VuetifyTemplate):
     """The primary table used in the ``DataManager`` class. This is a heavily modified vuetify
     DataTable, largely necessary for us to be able to put a more comprehensive set of action
     buttons in each row."""
+
+    template_file = vue_template_path("manager-table.vue")
 
     items = traitlets.List(traitlets.Dict()).tag(sync=True)
     headers = traitlets.List(traitlets.Dict()).tag(
@@ -111,96 +113,3 @@ class TableContentsTemplate(v.VuetifyTemplate):
     def vue_addToSample(self, point_id):
         for callback in self._add_to_sample_callbacks:
             callback(point_id)
-
-    # NOTE: I'm leaving this here for reference, I don't actually think there's any reason to need to go this direction
-    # but this is a functioning way to do it (note that this will _retrigger_ the updateOptions handler)
-    # def change_page(self, page_num):
-    #     new_options = dict(**self.options)
-    #     new_options["page"] = page_num
-    #     self.options = new_options
-
-    @traitlets.default("template")
-    def _template(self):
-        return """
-        <v-data-table
-            class='softhover-table'
-            :height="height"
-            :width="width"
-            dense
-            :items="items"
-            :headers="headers"
-            :server-items-length="total_length"
-            :options="options"
-            @update:options="updateOptions"
-        >
-            <template #body="{ items }">
-                <tbody :width="width">
-                    <tr v-for="item in items" :key="item.id" @click="selectPoint(item.id)" @mouseover="hoverPoint(item.id)">
-                        <td v-html="item.text" style="padding-left: 2px; padding-right: 2px; word-break: break-word;" />
-                        <td style="vertical-align: top; padding-left: 2px; padding-right: 2px; color: grey;">{{ item.id }}</td>
-                        <td style="vertical-align: top; padding-bottom: 5px; padding-left: 2px;">
-                            <v-tooltip bottom open-delay=500>
-                                <template v-slot:activator="{ on, attrs }">
-                                    <v-btn x-small class="blue darken-1" @click.stop="applyAbsoluteLabelUninteresting(item.id)" v-bind="attrs" v-on="on">
-                                        U
-                                    </v-btn>
-                                </template>
-                                <span>Label this item as <span class="blue--text lighten-4"><b>uninteresting</b></span> (cold).</span>
-                            </v-tooltip>
-                            <v-tooltip bottom open-delay=500>
-                                <template v-slot:activator="{ on, attrs }">
-                                    <v-btn x-small class="orange darken-1" @click.stop="applyAbsoluteLabelInteresting(item.id)" v-bind="attrs" v-on="on">
-                                        I
-                                    </v-btn>
-                                </template>
-                                <span>Label this item as <span class="orange--text lighten-5"><b>interesting</b></span> (warm).</span>
-                            </v-tooltip>
-                            <v-tooltip bottom open-delay=500 v-if="example_btn_color != ''">
-                                <template v-slot:activator="{ on, attrs }">
-                                    <v-btn x-small :style="{ backgroundColor: example_btn_color }" @click.stop="addToExampleAnchor(item.id)" v-bind="attrs" v-on="on">
-                                        example
-                                    </v-btn>
-                                </template>
-                                <span>Create a {{ example_type_name }} anchor with this item as the target.</span>
-                            </v-tooltip>
-                            <v-tooltip bottom open-delay=500>
-                                <template v-slot:activator="{ on, attrs }">
[…]
