+++ pyproject.toml
@@ -5,5 +5,6 @@ profile = "black"
 markers = [
     "integration",
     "vue",
-    "panel"
+    "panel",
+    "playwright"
 ]

[…]
