### 2024-06-21 155ba08
Add playwright trace upload to github test action
