### 2024-06-21 18c443a
Set github action to upload playwright trace if previous step fails
