+++ .github/workflows/tests.yml
@@ -27,6 +27,7 @@ jobs:
       - name: Test with pytest
         run: pytest -vv -s --tracing retain-on-failure
       - name: Archive failed playwright traces
+        if: failure()
         uses: actions/upload-artifact@v4
         with:
           name: traces
