### 2024-06-21 a0001f3
Add more browser delays in test
