+++ tests/test_browser.py
@@ -14,17 +14,17 @@ def test_jupyter(jupyter_server, page):
 
     # page.locator("div#jp-main-dock-panel > div.lm-Widget.jp-Editor.jp-InputArea-editor").first.click()
     page.get_by_label("Code Cell Content", exact=True).first.click()
-    page.keyboard.press("Shift+Enter")
-    page.keyboard.press("Shift+Enter")
-    page.keyboard.press("Shift+Enter")
-    page.keyboard.press("Shift+Enter")
-    page.keyboard.press("Shift+Enter")
-    page.keyboard.press("Shift+Enter")
-    page.keyboard.press("Shift+Enter")
-    page.keyboard.press("Shift+Enter")
+    page.keyboard.press("Shift+Enter", delay=1000)
+    page.keyboard.press("Shift+Enter", delay=1000)
+    page.keyboard.press("Shift+Enter", delay=1000)
+    page.keyboard.press("Shift+Enter", delay=1000)
+    page.keyboard.press("Shift+Enter", delay=1000)
+    page.keyboard.press("Shift+Enter", delay=1000)
+    page.keyboard.press("Shift+Enter", delay=1000)
+    page.keyboard.press("Shift+Enter", delay=1000)
 
     page.mouse.wheel(0, 1000)
-    expect(page.locator(".anchorviz")).to_be_visible(timeout=60000)
+    expect(page.locator(".anchorviz")).to_be_visible(timeout=240000)
 
     page.get_by_role("button", name="Dictionary").click()
 
