### 2024-06-21 0adb79a
Fix dependencies in github tests action
