+++ .github/workflows/tests.yml
@@ -24,7 +24,7 @@ jobs:
           python -m pip install --upgrade pip
           #pip install -r requirements.txt
           pip install -e .
-          pip install pytest-playwright jupyterlab
+          pip install pytest pytest-mock pytest-playwright jupyterlab
           playwright install
       - name: Test with pytest
         run: pytest -vv -s --tracing retain-on-failure
