+++ requirements.txt
@@ -2,8 +2,9 @@ panel
 numpy
 pandas
 scikit-learn
-ipyvuetify
-ipywidgets
+ipyvuetify~=1.9
+ipywidgets<8.1.3
+jupyterlab-widgets<3.0.11
 ipyanchorviz
 altair
 pytest
