+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.7.3] - 2024-06-21
+
+### Fixed
+* Missing `<template>` tags in vue templates
+* Overly loose dependencies in setup.py, restricted ipywidgets and ipyvuetify to
+  recent versions.
+
+
+
+
 ## [0.7.2] - 2023-11-29
 
 ### Fixed
