### 2024-06-21 ORNL/icat#44 @WarmCyan
I found and fixed a related bug that had started presenting with a recent `ipyvue` update as well, I believe `setup.py` in the updated 0.7.3 version now correctly reflects everything needed to get a currently working environment for JupyterLab

@jhagerer in theory you should now be able to run
```bash
git clone https://github.com/ORNL/icat.git
cd icat
conda create -n icat python
conda activate icat
conda install pip
pip install icat-iml
```
and have a functioning environment, let me know if it works for you and I'll close the issue
