### 2024-06-22 #6873 @jhrinv
Hi all, I am in contact with the authors. They have been actively working to fix some issues in the repository (broken package dependencies die to updates). I'll investigate the fixes and latest end of next week. I provide my checklist here then too. Thanks for your patience.
