### 2024-07-07 #6873 @jhagerer
Dear @WarmCyan , I was able to run the installation successfully. Thank you for the improvements! However, I am getting stuck in the execution of ICAT when I run `icat.initialize(offline=True)`, which is given by `notebooks/simple_example.ipynb` from the repository:

![image](https://github.com/openjournals/joss-reviews/assets/875050/d4e49c16-813b-45bd-ac7d-238c862db8dd)

Here you can inspect the full error message:
[error_stack.txt](https://github.com/user-attachments/files/16118916/error_stack.txt)

The same exception occurs when I run `icat.initialize(offline=True)` given by `lm_similarity_example.ipynb`.

What would be your remark on this? Thanks for your time and effort.
