### 2024-07-07 ORNL/icat#44 @jhagerer
Dear @WarmCyan , I was able to run the installation successfully. Thank you for the improvements! However, I am getting stuck in the execution of ICAT when I run `icat.initialize(offline=True)`, which is given by `notebooks/simple_example.ipynb` from the repository:

![Code_Q1YvXzFCnB](https://github.com/ORNL/icat/assets/875050/957464d8-e24e-499a-8e1a-aac832fc4d2b)

Here you can inspect the full error message:
[error_stack.txt](https://github.com/user-attachments/files/16118922/error_stack.txt)

The same exception occurs when I run `icat.initialize(offline=True)` given by `lm_similarity_example.ipynb`.

What would be your remark on this? Thanks for your time and effort.
