### 2024-07-07 ORNL/icat#44 @jhagerer
PS: I am using Python 3.12.4. Attached you see my installed packages.
[packages.txt](https://github.com/user-attachments/files/16119169/packages.txt)
