### 2024-07-07 ORNL/icat#44 @jhagerer
PPS: I just tried executing it as described before, but this time using Python 3.10.0. I receive the exact same error message `MultipleInstanceError: An incompatible sibling of 'Kernel' is already instantiated as singleton: str`.
