### 2024-07-16 ORNL/icat#44 @WarmCyan
I think this has to do with running it in vscode, this is likely related to a separate issue (about ipywidgets inside panel inside vscode) posted on the panel repo: https://github.com/holoviz/panel/issues/6175

Have you tried running this directly in Jupyter Lab? There are unfortunately compatibility issues sometimes when running widget-based stuff within vscode, so I'm only explicitly trying to support Jupyter Lab (I can add a note about this in the docs/README)
