### 2024-08-20 #6873 @WarmCyan
Hey @JBorrow! We've determined in https://github.com/ORNL/icat/issues/44 that ICAT doesn't run in vscode (which is somewhat expected/outside of what I'm trying to support.) I'm not sure if @jhagerer has had a chance to take a look at it since, but I'm waiting to see if it runs correctly for him in Jupyter Lab.
