### 2024-08-20 ORNL/icat#44 @jhagerer
@WarmCyan With Jupyter Lab and Jupyter Notebook, I am able to execute the provided simple_example.ipynb. According to the manual

https://ornl.github.io/icat/stable/user/interface.html

there should be some sort of GUI appearing. However, there is nothing showing up at all:

![image](https://github.com/user-attachments/assets/062e6490-a13e-4c69-96ba-0b0770f0f1c5)

I would have expected the same behaviour as described in the tutorial. What am I doing wrong here?
