### 2024-08-20 ORNL/icat#44 @jhagerer
@WarmCyan Forget my last post (I just removed it, since I had another problem setting up my conda environment). I got icat running successfully. I still have some feedback regarding usability which IMHO should be addressed. I will document my feedback over [here](https://github.com/openjournals/joss-reviews/issues/6873). Feel free to close the current issue.
