### 2024-08-20 #6873 @jhagerer
@WarmCyan I got icat running successfully. However, it looks honestly not fully usable, since I cannot recognize the text on the right, and also the illustration could be improved:

![image](https://github.com/user-attachments/assets/c3559f45-7a33-45dc-bd5f-25a768ace92a)

As you see, I used Jupyter Notebook and Jupyter Lab. Both look not ideal, even though both are recommended in the [official documentation](https://ornl.github.io/icat/stable/user/interface.html). What is your comment on that with regards to usability and correct display?

Further, I think it would be good to give a hint somewhere in the documentation that Visual Studio Jupyter code cells should not be used. Last but not least, I would recommend an advanced example that does not require transformers with a CUDA/NVIDIA hardware, but maybe something more lightweight but still reasonable and computationally efficient, e.g., word2vec.
