### 2024-08-21 ORNL/icat#45 @WarmCyan
JOSS review: Clean up interface to work in light mode and fit within Notebook
As pointed out by @jhagerer in https://github.com/openjournals/joss-reviews/issues/6873#issuecomment-2299852874, the interface doesn't render effectively in a light mode themed Jupyter instance, and is squished within Jupyter Notebook cells (as opposed to Jupyter Lab.)

Copying the example image over here for reference:
![image](https://github.com/user-attachments/assets/fefef736-c04c-40cf-a174-8164d29e2ebd)
