### 2024-08-21 #6873 @WarmCyan
@jhagerer yeah this is a good point, I didn't do due diligence in making sure it looks right in light mode as well, and that certainly has a huge impact on usability. The jarring mismatch of colors and inability to read the text should be fixable by checking for light/dark theme and adjusting interface coloring accordingly. To address the scrolling/crampedness of the jupyter notebook cells, I'll add a check for notebook vs lab and auto-scale components to make sure they fit.

I've opened an issue to track this here: https://github.com/ORNL/icat/issues/45 and will try to make both of these changes within the next week or two.
