### 2024-08-21 ORNL/icat#44 @WarmCyan
Thanks for the update! Going to leave this issue open and note for myself that I'll close this once I add to the README/docs that this doesn't work in VSCode
