### 2024-08-21 ORNL/icat#46 @WarmCyan
JOSS review: include intermediate example notebook
Added from @jhagerer https://github.com/openjournals/joss-reviews/issues/6873#issuecomment-2299852874: We go from a trivial "simple_example" notebook that's just expecting default anchors all the way to "lm_similarity_example" which adds a BERT transformer-based anchor (which has quite a few dependencies to get working.) It may be helpful to include an intermediate example that shows adding a custom anchor but doesn't require transformers to work. (word2vec was suggested as a possibility.)
