### 2024-08-21 #6873 @WarmCyan
That's also a good point about an additional example notebook that doesn't need transformers to run, I've opened an issue for it here: https://github.com/ORNL/icat/issues/46
