### 2024-08-28 1c93f2c
Improve makefile and clean up test
