### 2024-08-31 ORNL/icat#45 @jhagerer
Hey @WarmCyan , in case to speed up the review process, an easy quick win could also be achieved as follows: Just change the documentation by telling that you need to activate dark theme for Jupyter Lab. You can refer to the following link:

https://stackoverflow.com/a/46219796/776847
