### 2024-08-31 ORNL/icat#46 @jhagerer
@WarmCyan I would be happy with any example which does not need cuda/nvidia and gives me as a reviewer more to see and explore than the simple example.

A quick win solution could be to just change [lm_similarity_example.ipynb](https://github.com/ORNL/icat/blob/main/notebooks/lm_similarity_example.ipynb) such that it does not use cuda anymore. How about that?
