### 2024-09-04 6ffb63c
Add light theme handling for anchorviz ring colors
