+++ requirements.txt
@@ -5,7 +5,7 @@ scikit-learn
 ipyvuetify~=1.9
 ipywidgets<8.1.3
 jupyterlab-widgets<3.0.11
-ipyanchorviz
+ipyanchorviz~=0.3.0
 altair
 pytest
 pytest-mock
