+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.7.4] - 2024-09-13
+
+### Fixed
+* Unreadable text in light-themed Jupyter due to static colorscheme
+
+
+
+
 ## [0.7.3] - 2024-06-21
 
 ### Fixed
