### 2024-09-13 ORNL/icat#45 @WarmCyan
As it turned out solving the theming issue was much easier than making it fit in Jupyter Notebook. The interface should now be usable in both light-theme and dark-theme Jupyter Lab:
![image](https://github.com/user-attachments/assets/e7196d04-7be2-4f42-9165-66f5c861f5c5)

Note that this required an update to our underlying [ipyanchorviz](https://github.com/ORNL/ipyanchorviz) package, so your environment will need to be updated with `pip install --upgrade ipyanchorviz`

I wasn't able to find a good solution for making it fit correctly inside a Notebook cell -- the space is heavily constrained and there's not an easy way of detecting or overriding this. For now, I'm going to update the documentation to state that this is intended for use only in Jupyter Lab
