### 2024-09-27 ORNL/icat#46 @WarmCyan
@jhagerer good suggestion, I modified that notebook to first check if cuda is available, and use cpu otherwise.
