### 2024-10-06 ORNL/icat#45 @jhagerer
I can confirm it is working. You can close the ticket.
