### 2024-10-06 ORNL/icat#45 @jhagerer
I can confirm that this issue is solved now. Please close.
