### 2024-10-06 ORNL/icat#46 @jhagerer
OK, I think I have something running now to work with.

Please close the ticket. I'll take some time next weekend to try out and understand the tool more in depth.
