### 2024-11-03 #6873 @jhagerer
@JBorrow My problem is that for several months now, I am reading the documentation and try to make the best out of it. However, I have to say that the documentation does not enable me to use the tool.

Previously, the problem was that no UI would be visible after following the documentation. Then, there was no example which could be executed without GPU. Now, I cannot figure out how to seed a model and how to trigger a round of predictions in the interface. The documentation does not say anything about this: https://ornl.github.io/icat/stable/user/interface.html.

My gut feeling is that my experience should not be the problem. But maybe another reviewer should check if it is difficult for them too? For me this is very time consuming, and I keep having problem making anything work here.
