### 2024-11-04 #6873 @WarmCyan
@jhagerer sorry we keep running into difficulties with this, I know that can be frustrating. The seeding/prediction process is discussed in the docs here: https://ornl.github.io/icat/stable/user/concepts.html#labelling (basically you have to click the 'I'/'U' labelling buttons on at least 10 instances, and then it will automatically start making predictions)
