### 2024-11-06 #6873 @JBorrow
@jhagerer that is very unfortunate; would you be able to point me to issues that you've raised in the target repository and I can take a look? Sometimes different hardware and software environments are hard to mesh, though of course at JOSS we aim for the software that we accept to be widely useable.

I have just attempted the homepage documentation example and I was able to successfully get a UI on my M2 Macbook Pro:

However, @WarmCyan, it would definitely be better if there was some information along with this example that tells you _what_ to do with it? Even if there is some simple takeaway point from this simple example about how the UI works. Right now I just have a square of points in the middle of a giant circle, and it's not clear where to go from there at all.
