### 2024-11-25 #6873 @JBorrow
Hi folks, just wondering what we need to get unblocked here: @jhagerer I understand that you were facing some problems getting set up. Were you able to successfully raise issues against the target package and have them be resolved?
