### 2024-12-02 #6873 @WarmCyan
Thanks for the feedback @SamHames, and best wishes!

> it would have saved me a bit of time if there was some signposting in the UI about what's left to do before the model is "seeded"

That's a good point, I'll add some status indicators for the underlying model to make next logical actions more clear

> Is the intended output of this tool the machine learning model itself, or is the output the filtered dataset

The intent was really either/both. In principle, as you're building up the model trying to explore and filter down a particular dataset, at the end you have both the filtered dataset as well as a model that can then be used on more data or similar problems down the line. (User-provided labels are separately stored in the model class from the active dataset, so you can trade out datasets and the model will still take all previously labeled data into account.)

Definitely heard though, getting a pretty strong signal from everyone that it's not clear how exactly to use this/where to go with it once it's running. I'm out on travel next week, but I'll make some time this week and when I get back to put together a start-to-finish example or two for going through the process on one dataset and then maybe applying the model on another. I'll ping here once that's ready!
