### 2024-12-02 ORNL/icat#47 @WarmCyan
JOSS review: model status indicators
There's no explicit indication anywhere for whether the model is seeded or how many remaining labeled points are needed for the model to start running
(make sure to include notification if only I/U points are currently labeled)
