### 2024-12-02 ORNL/icat#48 @WarmCyan
JOSS review: End-to-end examples
The examples and documentation don't make it clear how to start using the tool once it's up and running.

Make one or two example notebooks demonstrating start to finish applications, at least one should include applying the model afterwards to a new dataset

(Also update the docs to provide more suggestions/info on next steps/what to actually do. Possibly a new page after Interface)
