### 2025-01-02 f2532d9
Fix error when switching data on already trained model
New dataset didn't have the predictions column yet which errored when
trying to serialize and send to anchorviz. Fixed by adding a
"data_changed" event that the model class listens for and calls `fit()`
