### 2025-01-02 ORNL/icat#48 @WarmCyan
A notebook demonstrating a full walk through (including applying the trained model to a new dataset) is now here: https://github.com/ORNL/icat/blob/main/notebooks/usage_walkthrough.ipynb
