### 2025-01-02 #6873 @WarmCyan
Whew, okay @SamHames @JBorrow @jhagerer sorry this took so long, I've put together a better tutorial/walkthrough notebook on basic strategies for how to use ICAT, in this example to pull out `sci.space` posts from 20newsgroups. Linked in the issue here: https://github.com/ORNL/icat/issues/48

I'll add some of the information and tips from this notebook into the documentation pages as well, and I still need to add a status indicator for model seeding, which I'll work on next!
