### 2025-01-14 922be02
Add status label to indicate conditions needed for seeding
