### 2025-01-16 faa1a08
Clean up status text
