### 2025-01-22 ORNL/icat#47 @WarmCyan
Still working on writing tests for this, but a basic text indicator has been added beneath the data manager widget, with text describing required conditions before model begins predicting
