### 2025-01-22 #6873 @WarmCyan
(To add, I haven't written tests yet/it's not yet in a new version release, but status text indicating requirements before model can predict has been added, for issue https://github.com/ORNL/icat/issues/47)
