### 2025-02-17 ORNL/icat#49 @SamHames
"Interesting" font colour is low contrast in light mode
Just a note that the orange font colour for interesting examples is low contrast and hard to read in light mode - it might be worth double checking all the colour choices have enough contrast for readability in both light and dark mode.
