### 2025-02-17 #6873 @SamHames
Sorry for the long delay - it turns out I broke my arm and it's taken a while to get back up to 100%.

I've reviewed the new tutorial and walkthrough and it makes the workflow and intended usage much clearer. I can sign off on  everything in the checklist except the section about community guidelines. That can be fixed with a couple of sentences in the readme (presumably issues can be reported on github?)

I have raised a minor issue about having enough contrast with coloured fonts, but that is in no way a blocker on this review. https://github.com/ORNL/icat/issues/49

I am happy to signoff on this once that minor note about the community guidelines is added somewhere to the readme.
