### 2025-02-24 1a13e8d
Add contributing section to README
