### 2025-02-24 #6873 @WarmCyan
I've added a section to the readme for contributing guidelines
