### 2025-05-09 #6873 @JBorrow
@WarmCyan looks like your bibfile is broken.
