### 2025-05-09 ORNL/icat#50 @JBorrow
JOSS Review: Functionality
A couple of functionality items that bothered me whilst using ICAT/I didn't understand:

- I see nothing in the docs about selection. I was able to select a bunch of items but they didn't seem to do anything - what's going on there?
- The software seems to be designed to work in jupyter but it doesn't fit nicely in a cell window on my computer independent of what browser scaling I used.

cc https://github.com/openjournals/joss-reviews/issues/6873
