### 2025-05-09 ORNL/icat#51 @JBorrow
JOSS Review: Documentation
- It would be good to have the statement of need and introduction from the README to be copied and pasted (or expanded upon) in the users guide.
- I would like to see the getting started page: https://ornl.github.io/icat/stable/user/getting_started.html include some information on adding some simple dictionary anchors based on news topics of the day and seeing what is closest to what. That was really cool to see in practice but the provided docs weren't really enough.

cc https://github.com/openjournals/joss-reviews/issues/6873
