### 2025-05-19 f073c12
Add basic usage section to getting started doc page
