### 2025-05-19 400d127
Add introduction section to user guide
