+++ sphinx/source/user_guide.rst
@@ -4,6 +4,7 @@ User Guide
 .. toctree::
     :maxdepth: 2
 
+    user/introduction.rst
     user/getting_started.rst
     user/concepts.rst
     user/interface.rst
