### 2025-05-23 5ef3eff
Add interface_width option and auto-detection for jupyter notebooks
