+++ setup.py
@@ -55,6 +55,7 @@ setup(
         "ipyvue~=1.11",
         "ipywidgets<8.1.3",  # can be removed once https://github.com/holoviz/panel/issues/6921 is resolved (panel 1.5?)
         "jupyterlab-widgets<3.0.11",  # "
+        "psutil",
     ],
     # an exact set of dependencies that I know for a fact works, remove comments
     # in case of emergency
