### 2025-05-23 f2f548d
Add explicit section in documentation on selection
