+++ sphinx/source/user/interface.rst
@@ -51,6 +51,9 @@ indicates that it was labeled by the user. Green indicates it is selected -
 points can be lasso selected by clicking and dragging on the inside of the ring
 and making a loop around a collection of points.
 
+.. figure:: ../_static/selection.png
+   :align: center
+
 This part of the interface can be rendered separately with
 ``model.view.anchorviz``.
 
