### 2025-05-23 cacefc8
Add notes to documentation on interface sizing
