+++ docs/0.8.0/_sources/api/icat.utils.vue_template_path.rst.txt
@@ -1,0 +1,6 @@
+icat.utils.vue\_template\_path
+==============================
+
+.. currentmodule:: icat.utils
+
+.. autofunction:: vue_template_path
