+++ docs/0.8.0/_sources/index.rst.txt
@@ -1,0 +1,31 @@
+.. ICAT documentation master file, created by
+   sphinx-quickstart on Wed Aug  2 09:22:31 2023.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to ICAT's documentation!
+================================
+
+.. toctree::
+    :maxdepth: 2
+    :caption: User Guide
+
+    user_guide.rst
+
+
+
+.. toctree::
+    :maxdepth: 2
+    :caption: API Reference
+
+    api_reference.rst
+
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
