+++ docs/0.8.0/_sources/api/icat.utils.jupyter_notebook_detected.rst.txt
@@ -1,0 +1,6 @@
+icat.utils.jupyter\_notebook\_detected
+======================================
+
+.. currentmodule:: icat.utils
+
+.. autofunction:: jupyter_notebook_detected
