+++ docs/0.8.0/_sources/api/icat.utils.jupyter_lab_detected.rst.txt
@@ -1,0 +1,6 @@
+icat.utils.jupyter\_lab\_detected
+=================================
+
+.. currentmodule:: icat.utils
+
+.. autofunction:: jupyter_lab_detected
