+++ docs/0.8.0/_sources/api/icat.model.Model.rst.txt
@@ -1,0 +1,45 @@
+icat.model.Model
+================
+
+.. currentmodule:: icat.model
+
+.. autoclass:: Model
+   :members:
+   :show-inheritance:
+   :inherited-members:
+
+
+
+
+   .. rubric:: Methods
+
+   .. autosummary::
+
+      ~Model.__init__
+      ~Model.add_anchor
+      ~Model.compute_coverage
+      ~Model.feature_names
+      ~Model.featurize
+      ~Model.fire_on_status_event
+      ~Model.fit
+      ~Model.is_seeded
+      ~Model.is_trained
+      ~Model.load
+      ~Model.on_status_event
+      ~Model.predict
+      ~Model.save
+
+
+
+
+
+   .. rubric:: Attributes
+
+   .. autosummary::
+
+      ~Model.training_data
+      ~Model.text_col
+      ~Model.classifier
+      ~Model.anchor_list
+      ~Model.data
+      ~Model.view
