+++ docs/0.8.0/_sources/api/icat.model.rst.txt
@@ -1,0 +1,22 @@
+﻿icat.model
+==========
+
+.. automodule:: icat.model
+
+
+
+
+
+
+
+
+
+
+
+   .. rubric:: Classes
+
+   .. autosummary::
+      :toctree:
+      :template: custom-class-template.rst
+
+      Model
