+++ docs/0.8.0/_sources/user_guide.rst.txt
@@ -1,0 +1,10 @@
+User Guide
+==========
+
+.. toctree::
+    :maxdepth: 2
+
+    user/introduction.rst
+    user/getting_started.rst
+    user/concepts.rst
+    user/interface.rst

[…]
