### 2025-05-23 961ceb5
Remove overly strict dependency versions
