### 2025-05-23 bba70ef
Remove strict dependencies in requirements.txt
