### 2025-05-23 ORNL/icat#45 @WarmCyan
Fitting within a notebook cell has since been fixed in https://github.com/ORNL/icat/commit/5ef3effe10f717150ec52417842c7ecf0181d9b8
