### 2025-05-23 ORNL/icat#51 @WarmCyan
> It would be good to have the statement of need and introduction from the README to be copied and pasted (or expanded upon) in the users guide.

I added an [Introduction](https://ornl.github.io/icat/stable/user/introduction.html) section to the user guide with those sections of the paper.

> I would like to see the getting started page: https://ornl.github.io/icat/stable/user/getting_started.html include some information on adding some simple dictionary anchors based on news topics of the day and seeing what is closest to what. That was really cool to see in practice but the provided docs weren't really enough.

Good point, I added a [Basic Usage](https://ornl.github.io/icat/stable/user/getting_started.html#basic-usage) section to the getting started page with enough instructions to actually see something/start using it, plus an explicit link to the walkthrough notebook.

Let me know if you think those are sufficient!
