### 2025-05-23 ORNL/icat#50 @WarmCyan
> I see nothing in the docs about selection. I was able to select a bunch of items but they didn't seem to do anything - what's going on there?

I had a couple very brief phrases mentioning it in passing in the interface section, but not at all obvious or particularly helpful, so I added some more explicit info on this in a new [Selection](https://ornl.github.io/icat/stable/user/concepts.html#selection) section of the concepts page.
(The main point of selection being to pull those specific points up in the "Selected" tab of the data table on the right - a tab that was very likely not even visible at first if the width was an issue!)

> The software seems to be designed to work in jupyter but it doesn't fit nicely in a cell window on my computer independent of what browser scaling I used.

Out of curiosity is this in Jupyter Notebook or is it Lab but just on a smaller screen? Fixing width has been long overdue, as was initially pointed out in https://github.com/ORNL/icat/issues/45, so I finally added a parameter to allow manually setting the width of the overall interface (with a note on how to change it in [this section](https://ornl.github.io/icat/stable/user/getting_started.html#user-interface) in the docs). In principle it should now also detect if you're running in Notebook and automatically set the width appropriately, but if it doesn't/you're in Lab but with a smaller screen, you can now manually make it smaller by passing `interface_width` to the `Model` init
