### 2025-05-23 62c7978
Fix missing comma in bibfile
