+++ paper/paper.bib
@@ -127,7 +127,7 @@ note={Accessed: 2023-10-12},
          and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
          and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
          Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
- doi = {10.48550/arXiv.1201.0490}
+ doi = {10.48550/arXiv.1201.0490},
  journal={Journal of Machine Learning Research},
  volume={12},
  pages={2825--2830},
