### 2025-05-23 #6873 @WarmCyan
Hey @JBorrow, thanks for the feedback! I think I've addressed all your points in the two issues, let me know if there's anything else you'd like to see or if I'm good to close those

I also fixed the missing comma issue in the bibfile
