### 2025-05-30 #6873 @JBorrow
Copy-edits:

Line 40 - please put these two citation references in the same parentheses (`[@amershiPowerPeopleRole2014, @cakmakDesigningInteractionsRobot2010]`, not `[@amershiPowerPeopleRole2014],[@cakmakDesigningInteractionsRobot2010]`)

Line 44 - you may wish to just say `interface concepts proposed by @suhAnchorVizFacilitatingSemantic2019` to render the citation without parentheses.

See the pandoc docs for clarification if necessary: https://pandoc.org/MANUAL.html#citations-in-note-styles
