### 2025-06-03 213b63a
Make suggested citation formatting changes to paper
