### 2025-06-03 #6873 @JBorrow
@WarmCyan please could you take a look at the 'Additional Author Tasks After Review is Complete'?

In particular, please make sure that the archive has the following:

- Exactly the same name as the paper
- Exactly the same authors and ORCIDs as the paper
- Exactly the same license as the software
