### 2025-06-09 85e9eaa
Bump version for JOSS release
