+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.8.1] - 2025-06-09
+
+Version incorporating all final changes to JOSS paper.
+
+
+
+
 ## [0.8.0] - 2025-05-23
 
 ### Added
