+++ environment.yml
@@ -10,7 +10,7 @@ dependencies:
 - scikit-learn
 - nodejs
 - pandas
-- ipyvuetify~=1.9
-- ipywidgets<8.1.3
+- ipyvuetify
+- ipywidgets
 - pytest
 - pytest-mock
