### 2025-06-09 #6873 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
