### 2025-06-10 e262033
Make requested paper edits
