+++ paper/paper.bib
@@ -105,7 +105,7 @@
 
 
 @misc{panelholoviz,
-title={Panel: The powerful data exploration \& web app framework for Python},
+title={Panel: The powerful data exploration \& web app framework for {P}ython},
 author={Holoviz},
 doi = {10.5281/zenodo.3706648},
 howpublished={\url{https://panel.holoviz.org/}},
