### 2025-06-10 #6873 @crvernon
🥳 Congratulations on your new publication @WarmCyan! Many thanks to @JBorrow for editing and @SamHames and @JBorrow (thanks for stepping in and getting this knocked out @JBorrow) for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
